+++ src/components/App/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {BrowserRouter, Routes, Route} from "react-router-dom";
 
-import {Slider} from '../Slider'
+import {Slider} from '../Slider/slider'
 
 import './style.scss';
