+++ src/components/Slider/MoviesPreview/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import {createItem} from '../index'
-import {IMoviesPreview} from "../../../types";
+import {createItem} from '../slider'
+import {IMoviesPreview} from "../types";
 
 import './style.scss'
 
@@ -12,6 +12,7 @@ export const MoviesPreview = ({pos, idx}: IMoviesPreview) => {
         <div className="preview__img">
           <div className="preview__info">
             <h3 className="title preview__title">{item.movie.title}</h3>
+            <span className="preview__duration">{item.movie.duration}</span>
             <button className="gen-button preview__button-genre">Action</button>
           </div>
           <img src={item.movie.image} alt={item.movie.title}/>
