--- src/components/Slider/style.scss
@@ -1,37 +1,0 @@
-@font-face {
-  font-family: 'Jost';
-  src: url("../../styles/fonts/Jost/Jost-Regular.ttf");
-}
-
-@font-face {
-  font-family: 'Roboto';
-  src: url("../../styles/fonts/Roboto/Roboto-Regular.ttf");
-}
-
-@font-face {
-  font-family: 'Roboto-Light';
-  src: url("../../styles/fonts/Roboto/Roboto-Light.ttf");
-}
-
-img {
-  width: 100%;
-}
-
-.title {
-  font-family: Jost, sans-serif;
-  line-height: 1;
-  color: #fff;
-}
-
-.gen-button {
-  cursor: pointer;
-  background-color: #e50916;
-  border: none;
-  color: #ffffff;
-  font-family: Jost, sans-serif;
-  font-size: 16px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 12px;
-}
