+++ src/components/Slider/types.ts
