--- src/components/Video/Video.scss
@@ -1,21 +1,0 @@
-.player {
-  display: flex;
-  justify-content: center;
-  position: relative;
-  overflow: hidden;
-  margin: 2% auto;
-  width: 95%;
-  max-height: 87vh;
-
-  &__container {
-    max-width: 95%;
-    max-height: 87vh;
-  }
-
-  &__button {
-    position: fixed;
-    top: 5%;
-    right: 5%;
-  }
-
-}
